// src/enums.ts
export enum PostType {
    NEW = "NEW",
    EXISTING = "EXISTING",
}

export enum StrategyType {
    SIMPLE = "SIMPLE",
    SMART = "SMART",
}

export enum TaskStatus {
    CREATED = "CREATED",
    STARTED = "STARTED",
    DONE = "DONE",
    ERROR = "ERROR",
}

export enum OrderStatus {
    CREATED = "CREATED",
    SENT = "SENT",
    DONE = "DONE",
    ERROR = "ERROR",
}